all done

+++ slave/analyzer/tools/basic_histogram.rb
@@ -2,29 +2,8 @@ class BasicHistogram < AnalysisMetadata
 
   DEFAULT_CHUNK_SIZE = 1000
   
-  def self.set_variables(analysis_metadata, curation)
-    remaining_variables = []
-    analysis_metadata.analytical_offering.variables.each do |variable|
-      analytical_offering_variable = AnalyticalOfferingVariable.new
-      analytical_offering_variable.analytical_offering_variable_descriptor_id = variable.id
-      analytical_offering_variable.analysis_metadata_id = analysis_metadata.id
-      case variable.name
-      when "curation_id"
-        analytical_offering_variable.value = curation.id
-        analytical_offering_variable.save
-      when "save_path"
-        analytical_offering_variable.value = "analytical_results/#{analysis_metadata.function}"
-        analytical_offering_variable.save
-      else
-        remaining_variables << variable
-      end
-    end
-    return remaining_variables
-  end
-
-  
   #Results: Frequency Charts of basic data on Tweets and Users per data set
-  def self.run(curation_id, save_path)
+  def self.run(curation_id)
     curation = Curation.first(:id => curation_id)
     FilePathing.tmp_folder(curation, self.underscore)
     self.generate_graphs([
@@ -53,12 +32,17 @@ class BasicHistogram < AnalysisMetadata
       fs[:style] = fs[:style] || "histogram"
       fs[:title] = fs[:title] || fs[:model].pluralize+"_"+fs[:attribute].to_s
       fs[:conditional] = fs[:conditional] || {}
-      graph_attrs = Hash[fs.select{|k,v| Graph.attributes.include?(k)}]
-      graph = Graph.first_or_create({:curation_id => curation_id, :analysis_metadata_id => analytic.analysis_metadata(curation).id}.merge(graph_attrs))
-      graph.graph_points.destroy #can't call .new? as a condition for this, as it's created now.
-      graph.edges.destroy #can't call .new? as a condition for this, as it's created now.
-      conditional = Analysis.curation_conditional(curation).merge(fs[:conditional])
-      graphs << graph
+      fs[:generate_graph_points] = fs[:generate_graph_points] || true
+      fs[:override_conditional] = fs[:override_conditional] || false
+      graph = nil
+      if fs[:generate_graph_points]
+        graph_attrs = Hash[fs.select{|k,v| Graph.attributes.include?(k)}]
+        graph = Graph.first_or_create({:curation_id => curation_id, :analysis_metadata_id => analytic.analysis_metadata(curation).id}.merge(graph_attrs))
+        graph.graph_points.destroy #can't call .new? as a condition for this, as it's created now.
+        graph.edges.destroy #can't call .new? as a condition for this, as it's created now.
+        conditional = fs[:override_conditional] ? fs[:conditional] : Analysis.curation_conditional(curation).merge(fs[:conditional])
+        graphs << graph
+      end
       if block_given?
         yield fs, graph, conditional
       else
@@ -72,22 +56,28 @@ class BasicHistogram < AnalysisMetadata
     limit = DEFAULT_CHUNK_SIZE||1000
     offset = 0
     sub_directory = "/"+[fs[:year],fs[:month],fs[:date],fs[:hour]].compact.join("/")
-    full_path_with_file = sub_directory == "/" ? path+"/"+graph.title+".csv" : path+sub_directory+"/"+graph.title+".csv"
+    full_path_with_file = sub_directory == "/" ? path+"/"+fs[:title]+".csv" : path+sub_directory+"/"+fs[:title]+".csv"    
     Sh::mkdir(path+sub_directory) if sub_directory != "/"
     FasterCSV.open(full_path_with_file, "w") do |csv|
-      records = fs[:model].aggregate(fs[:attribute], :all.count, {:limit => limit, :offset => offset}.merge(conditional))
-      graph_points = records.collect{|record| {:label => record.first, :value => record.last, :graph_id => graph.id, :curation_id => graph.curation_id}}
-      graph_points = graph.sanitize_points(graph_points)
-      while !records.empty?
-        csv << ["label", "value"]
-        graph_points.each do |graph_point|
-          csv << [graph_point[:label],graph_point[:value]]
-        end
-        GraphPoint.save_all(graph_points)
-        offset+=limit
+      if block_given?
+        yield fs, graph, conditional, csv, limit, offset
+      else
         records = fs[:model].aggregate(fs[:attribute], :all.count, {:limit => limit, :offset => offset}.merge(conditional))
+        while !records.empty?
+          graph_points = records.collect{|record| {:label => record.first, :value => record.last, :graph_id => graph.id, :curation_id => graph.curation_id}}
+          graph_points = graph.sanitize_points(graph_points)
+          GraphPoint.save_all(graph_points) if fs[:generate_graph_points]
+          csv << ["label", "value"]
+          graph_points.each do |graph_point|
+            csv << [graph_point[:label],graph_point[:value]]
+          end
+          offset+=limit
+          records = fs[:model].aggregate(fs[:attribute], :all.count, {:limit => limit, :offset => offset}.merge(conditional))
+        end
       end
     end
+    graph.written = true
+    graph.save!
   end
 
   def self.finalize_analysis(curation)
